refactor: 게임 메세지 출력 방식 리팩터링
