+++ src/main/java/baseball/BaseballGame.java
@@ -2,31 +2,32 @@ package baseball;
 
 import baseball.console.GameConsole;
 import baseball.converter.ThreeNumberConverter;
+import baseball.dto.BallStrikeDto;
 import baseball.opponent.ComputerOpponent;
-import baseball.printer.MessagePrinter;
-import baseball.printer.NumberInputMessagePrinter;
-import baseball.printer.StartMessagePrinter;
+import baseball.opponent.RandomNumberGenerator;
+import baseball.printer.GameMessagePrinter;
 
 import java.util.List;
 
 public class BaseballGame {
 
     private ComputerOpponent opponent;
-    private MessagePrinter inputMessagePrinter;
     private ThreeNumberConverter threeNumberConverter;
+    private GameMessagePrinter messagePrinter;
     private GameConsole console;
 
     public BaseballGame() {
-        MessagePrinter startMessagePrinter = new StartMessagePrinter();
-        this.inputMessagePrinter = new NumberInputMessagePrinter();
+        this.messagePrinter = new GameMessagePrinter();
         this.opponent = new ComputerOpponent(RandomNumberGenerator.generate());
-        startMessagePrinter.printMessage();
+
+        messagePrinter.printStartMessage();
     }
 
     public void play() throws IllegalArgumentException {
         try {
-            inputMessagePrinter.printMessage();
+            messagePrinter.printNumberInputMessage();
             List<Integer> threeNumbers = threeNumberConverter.getThreeNumbers(console.input());
+            BallStrikeDto answer = opponent.answer(threeNumbers);
         } catch (Exception e) {
             throw new IllegalArgumentException(e.getMessage());
         }
