--- src/main/java/baseball/printer/MessagePrinter.java
@@ -1,5 +1,0 @@
-package baseball.printer;
-
-public interface MessagePrinter {
-    void printMessage();
-}
